Create Game State Enum

+++ Space-Invaders/Source/Time/TimeService.cpp
@@ -25,8 +25,8 @@ namespace Time {
 	float TimeService::calculateDeltaTime()
 	{
 		// Calculate time difference in microseconds between the current and previous frame.
-		int delta = std::chrono::duration_cast<std::chrono::microseconds>(
-			std::chrono::steady_clock::now() - previous_time).count();
+		int delta = static_cast<int>(std::chrono::duration_cast<std::chrono::microseconds>(
+			std::chrono::steady_clock::now() - previous_time).count());
 
 		// The cast is used to convert delta time from microseconds into seconds.
 		// We will learn aboit how this works in detail later.
